+++ public/templates/sidebar_home.inc.php
@@ -34,12 +34,17 @@ use Ampache\Module\System\Core;
 use Ampache\Module\Util\Ui;
 use Ampache\Repository\VideoRepositoryInterface;
 
+/** @var string $t_browse */
+/** @var string $t_expander */
+
 global $dic;
 $server_allow    = AmpConfig::get('allow_localplay_playback');
 $controller      = AmpConfig::get('localplay_controller');
 $videoRepository = $dic->get(VideoRepositoryInterface::class);
 $allowVideo      = AmpConfig::get('allow_video') && $videoRepository->getItemCount(Video::class);
 $allowDemocratic = AmpConfig::get('allow_democratic_playback');
+$showAlbumArtist = AmpConfig::get('show_album_artist');
+$showArtist      = AmpConfig::get('show_artist');
 $allowLabel      = AmpConfig::get('label');
 $allowPodcast    = AmpConfig::get('podcast');
 $access50        = Access::check('interface', 50);
@@ -60,7 +65,12 @@ $web_path        = AmpConfig::get('web_path'); ?>
         <ul class="sb3" id="sb_browse_music">
             <li id="sb_home_browse_music_songTitle"><a href="<?php echo $web_path; ?>/browse.php?action=song"><?php echo $t_songs ?></a></li>
             <li id="sb_home_browse_music_album"><a href="<?php echo $web_path; ?>/browse.php?action=album"><?php echo $t_albums; ?></a></li>
-            <li id="sb_home_browse_music_artist"><a href="<?php echo $web_path; ?>/browse.php?action=album_artist"><?php echo $t_a_artists; ?></a></li>
+            <?php if ($showAlbumArtist || (!$showArtist)) { ?>
+                <li id="sb_home_browse_music_artist"><a href="<?php echo $web_path; ?>/browse.php?action=album_artist"><?php echo $t_a_artists; ?></a></li>
+            <?php } ?>
+            <?php if ($showArtist) { ?>
+                <li id="sb_home_browse_music_artist"><a href="<?php echo $web_path; ?>/browse.php?action=artist"><?php echo $t_artists; ?></a></li>
+            <?php } ?>
             <?php if ($allowLabel) { ?>
                 <li id="sb_home_browse_music_label"><a href="<?php echo $web_path; ?>/browse.php?action=label"><?php echo $t_labels ?></a></li>
             <?php } ?>